backend/requirements.txt
# flask==3.0.3
# flask-cors==4.0.1
# anthropic==0.34.1
# requests==2.32.3
# python-dotenv==1.0.1
# gunicorn==22.0.0

flask==3.0.3
flask-cors==4.0.1
requests==2.32.3
python-dotenv==1.0.1
gunicorn==22.0.0
groq==0.9.0
duckduckgo-search==6.2.6